attempted to fix move from tutorial to the rooms
I know what's going on but don't know how to fix it. The issue is trying to pass a reference to the appropriate room properly in the goNorth, etc functions in main and then calling the function to load that room. Note sections of rooms and game (the flag variable entry and the end tutorial function) as well if attempting to fix.